Allow second parameter on review command
This allows to specify a tag different from the current branch:
git review <branch> <tag>

If tag is not specified the name of the current branch will be used.

Example:
if we're currently working in the ABC branch:
git review develop

The change will be listed as "develop (ABC)" in Gerrit.

git review develop XYZ
The change will be listed as "develop (XYZ)" in Gerrit.