Add hook that checks commit message
The commit message should follow the best practices: first line
is the subject and should be less than 62 characters, not ending
with period. Second line should be empty. Following lines should
be less than 75 chars. The hook verifies that this conditions are
met and prevents a commit otherwise.

This change also contains a wrapper script that allows to call
multiple commit-msg hooks, the new one above as well as the one
that gerrit provides.

Change-Id: Ib902aad3ef23135016203cfac36886a1945205cd