Set CRLF conversion on Windows
On Windows we need to set core.autocrlf to true. This is the default
on most systems, but on at least one machine it was set to false,
causing whitespace errors when trying to commit. Besides we don't want
to submit any \r\n in text files because this will cause build problems
on Linux.

Change-Id: I96200bf11283162038bded9746b6bf8d0ca1a72a