Improve previous commit
We still want to read the values from stdin even when run on a
dumb terminal. This is necessary so that desired values can be
piped in. The difference why we need this (and/or the previous
change) is that the packaging agents don't have /dev/tty - maybe
because they are based on Ubuntu Server instead of Desktop now.

Change-Id: Iaae6267e318f5ae8a3249bcc3a16e723ad024628